refactor : Coin 관련 리팩토링
