+++ src/main/java/vendingmachine/controller/OrderController.java
@@ -1,7 +1,7 @@
 package vendingmachine.controller;
 
 import vendingmachine.domain.coin.Coin;
-import vendingmachine.domain.coin.util.CoinRandomProvider;
+import vendingmachine.domain.coin.provider.CoinRandomProvider;
 import vendingmachine.domain.consumer.Consumer;
 import vendingmachine.domain.machine.VendingMachine;
 import vendingmachine.util.IOProvider;
