+++ src/main/java/vendingmachine/domain/coin/provider/CoinProvider.java
@@ -1,4 +1,4 @@
-package vendingmachine.domain.coin.util;
+package vendingmachine.domain.coin.provider;
 
 public interface CoinProvider {
     int drawCoinLessThanBalance(int balance);
